r/network_interface: conditionally updating the nics only for ipv4 configurations
